Add NYC taxi data dump
We use the NYC taxi data in our workshops. This db dump was in the GX
Core repository, but we remvoed it for the 1.0 release along with
other example data. In order to preserve its availability and use in
our workshop materials, let's copy it here where we use it in the
Dockerfile.

Note that we're not using Docker containers with workshops since
changes to the GX Cloud backend in July. However, we may still want to
use this data and the Dockerfile in the future or with other data
sources added in the future.