[DSB-1138] add databricks sql workshop
This commit adds the Databricks SQL workshop. Some of the images that
are "common" are included because I saw that the data source name is
in the corner, and I wanted to have that line up with the correct data
source for the workshop so that users have consistency in what they're
seeing for this one. We can redo these for the other workshops, as
we'll likely need to update those soon for other UI changes in GX
Cloud anyway.

This also has an updated "`new_passenger_expectation.png`" image
because we encountered an issue in a recent live workshop where not
putting in the `Min` value resulted in an exception.